fixed ray cast

+++ src/geom.cpp
@@ -21,7 +21,7 @@ bool onSegment(Point p, Point q, Point r){
 // 1 --> Clockwise
 // 2 --> Counterclockwise
 int orientation(Point p, Point q, Point r){
-    int val = (q.y - p.y) * (r.x - q.x) -
+    double val = (q.y - p.y) * (r.x - q.x) -
               (q.x - p.x) * (r.y - q.y);
  
     if (val == 0) return 0;  // colinear
